requirements.txt
pydocumentdb==2.3.2
python-dotenv==0.8.2
redis==2.10.6
requests==2.18.4
